requirements.txt
numpy~=2.1.3
tensorflow~=2.19.0
opencv-python~=4.11.0.86
six~=1.17.0
uvicorn~=0.34.0
fastapi~=0.115.12
scikit-learn~=1.6.1
scipy~=1.15.2
imageio~=2.37.0
mysql-connector-python~=9.3.0
starlette~=0.46.1
python-multipart
pillow~=11.2.1
pydantic~=2.11.3
pyvi~=0.1.1
regex~=2024.11.6
sentence-transformers~=4.1.0
tf-keras~=2.19.0